Update Range Assignment Q3.py

+++ 4.9.25/Range Assignment Q3.py
@@ -2,7 +2,7 @@
 #Date: 4/9/25
 #Description: 3. Ask the user to enter a number between 10 and 20 inclusive. If they enter a number in this range display the 'Thank you.'. Otherwise, display the message 'incorrect value entered.'.
 
-number = int(input('Enter a number between 10 and 20: '))
+number = float(input('Enter a number between 10 and 20: '))
 
 if (number > 10) and (number < 20):
     print(f'{number} is in the range.')
@@ -12,3 +12,4 @@ if(number> 20) or (number < 10):
     
 else:
     print(f'{number} is in the range.')
+
